refactor of stochastic PYCUDA engines (EPIE/SDR) - tests passing
