+++ ptypy/accelerate/cuda_pycuda/cuda/ob_update_local.cu
@@ -33,8 +33,9 @@ extern "C" __global__ void ob_update_local(
     int H,
     int I,
     const int* __restrict__ addr,
-    const IN_TYPE* __restrict__ pr_norm_sum,
-    IN_TYPE step)
+    const IN_TYPE* pr_norm_max,
+    const IN_TYPE A_,
+    const IN_TYPE B_)
 {
   const int bid = blockIdx.z;
   const int tx = threadIdx.x;
@@ -50,8 +51,9 @@ extern "C" __global__ void ob_update_local(
   probe += pa[0] * E * F + pa[1] * F + pa[2];
   obj += oa[0] * H * I + oa[1] * I + oa[2];
   aux += bid * B * C;
-  MATH_TYPE max_norm_val = pr_norm[0];
-  const MATH_TYPE step_val = step;
+  const MATH_TYPE pr_norm_max_val = pr_norm_max[0];
+  const MATH_TYPE A_val = A_;
+  const MATH_TYPE B_val = B_;
   
   assert(oa[0] * H * I + oa[1] * I + oa[2] + (B - 1) * I + C - 1 < G * H * I);
 
@@ -62,9 +64,9 @@ extern "C" __global__ void ob_update_local(
       complex<MATH_TYPE> probe_val = probe[b * F + c];
       complex<MATH_TYPE> exit_val = exit_wave[b * C + c];
       complex<MATH_TYPE> aux_val = aux[b * C + c];
-      MATH_TYPE norm_val = (MATH_TYPE(1) - step_val) * max_norm_val + step_val * pr_norm_sum[b * F + c];
-      
-      auto add_val_m = conj(probe_val) * (exit_val - aux_val) / norm_val;
+      MATH_TYPE norm_val = (MATH_TYPE(1) - A_val) * pr_norm_max_val + A_val * pr_norm[b * F + c];
+
+      auto add_val_m = (A_val + B_val) * conj(probe_val) * (exit_val - aux_val) / norm_val;
       complex<OUT_TYPE> add_val = add_val_m;
       atomicAdd(&obj[b * I + c], add_val);
   }
