+++ ptypy/accelerate/cuda_pycuda/kernels.py
@@ -1018,28 +1018,14 @@ class PoUpdateKernel(ab.PoUpdateKernel):
                                  block=(16, 16, 1), grid=grid, stream=self.queue)
 
 
-    def ob_update_local(self, addr, ob, pr, ex, aux, step):
-        # lazy allocation of temporary 1-element array
-        if self.norm is None:
-            self.norm = gpuarray.empty((1,), dtype=np.float32)
-        self.MAK.max_abs2(pr, self.norm)
-        
-        # TODO: turn this into proper CUDA kernel
-        from pycuda.elementwise import ElementwiseKernel
-        from reikna.algorithms import Reduce, predicate_sum
-        pr_power = pr.copy().__abs__()
-        print("GPU: ", pr.get())
-        print("GPU: ", pr_power.get())
-        pr_power_sum = gpuarray.zeros((pr.shape[-2], pr.shape[-1]), dtype=np.float32)
-        Reduce(pr_power, predicate_sum(np.float32), [0], pr_power_sum)
-        print("GPU: ", pr_power_sum.get())
-
+    def ob_update_local(self, addr, ob, pr, ex, aux, prn, A=0., B=1.):
+        prn_max = gpuarray.max(prn, stream=self.queue)
         obsh = [np.int32(ax) for ax in ob.shape]
         prsh = [np.int32(ax) for ax in pr.shape]
         exsh = [np.int32(ax) for ax in ex.shape]
         # atomics version only
         if addr.shape[3] != 3 or addr.shape[2] != 5:
-            raise ValueError('Address not in required shape for tiled pr_update')
+            raise ValueError('Address not in required shape for tiled ob_update')
         num_pods = np.int32(addr.shape[0] * addr.shape[1])
         bx = 64
         by = 1
@@ -1047,22 +1033,19 @@ class PoUpdateKernel(ab.PoUpdateKernel):
             exsh[0], exsh[1], exsh[2],
             pr,
             prsh[0], prsh[1], prsh[2],
-            self.norm,
+            prn,
             ob,
             obsh[0], obsh[1], obsh[2],
             addr,
-            pr_power_sum,
-            np.float32(step),
+            prn_max,
+            np.float32(A),
+            np.float32(B),
             block=(bx, by, 1),
             grid=(1, int((exsh[1] + by - 1)//by), int(num_pods)),
             stream=self.queue)
 
-    def pr_update_local(self, addr, pr, ob, ex, aux, step):
-        # lazy allocation of temporary 1-element array
-        if self.norm is None:
-            self.norm = gpuarray.empty((1,), dtype=np.float32)
-        self.MAK.max_abs2(ob, self.norm)
-        
+    def pr_update_local(self, addr, pr, ob, ex, aux, obn, A=0., B=1.):
+        obn_max = gpuarray.max(obn, stream=self.queue)
         obsh = [np.int32(ax) for ax in ob.shape]
         prsh = [np.int32(ax) for ax in pr.shape]
         exsh = [np.int32(ax) for ax in ex.shape]
@@ -1070,17 +1053,19 @@ class PoUpdateKernel(ab.PoUpdateKernel):
         if addr.shape[3] != 3 or addr.shape[2] != 5:
             raise ValueError('Address not in required shape for tiled pr_update')
         num_pods = np.int32(addr.shape[0] * addr.shape[1])
-
         bx = 64
         by = 1
         self.pr_update_local_cuda(ex, aux,
             exsh[0], exsh[1], exsh[2],
             pr,
             prsh[0], prsh[1], prsh[2],
-            self.norm,
+            obn,
             ob,
             obsh[0], obsh[1], obsh[2],
             addr,
+            obn_max,
+            np.float32(A),
+            np.float32(B),
             block=(bx, by, 1),
             grid=(1, int((exsh[1] + by - 1) // by), int(num_pods)),
             stream=self.queue)
