+++ test/accelerate_tests/cuda_pycuda_tests/po_update_kernel_test.py
@@ -670,8 +670,6 @@ class PoUpdateKernelTest(PyCudaTest):
         total_number_modes = G * D
         A = total_number_scan_positions * total_number_modes # this is a 16 point scan pattern (4x4 grid) over all the modes
 
-        step = 0.1
-
         probe = np.empty(shape=(D, E, F), dtype=COMPLEX_TYPE)
         for idx in range(D):
             probe[idx] = np.ones((E, F)) * (idx + 1) + 1j * np.ones((E, F)) * (idx + 1)
@@ -684,6 +682,8 @@ class PoUpdateKernelTest(PyCudaTest):
         for idx in range(A):
             exit_wave[idx] = np.ones((B, C)) * (idx + 1) + 1j * np.ones((B, C)) * (idx + 1)
         auxiliary_wave = exit_wave.copy() * 2
+
+        probe_norm = np.empty(shape=(1,B,C), dtype=FLOAT_TYPE)
 
         X, Y = np.meshgrid(range(scan_pts), range(scan_pts))
         X = X.reshape((total_number_scan_positions))
@@ -717,10 +717,13 @@ class PoUpdateKernelTest(PyCudaTest):
         probe_dev = gpuarray.to_gpu(probe)
         exit_wave_dev = gpuarray.to_gpu(exit_wave)
         auxiliary_wave_dev = gpuarray.to_gpu(auxiliary_wave)
+        probe_norm_dev = gpuarray.to_gpu(probe_norm)
         addr_dev = gpuarray.to_gpu(addr)
 
-        POUK.ob_update_local(addr_dev, object_array_dev, probe_dev, exit_wave_dev, auxiliary_wave_dev, step)
-        nPOUK.ob_update_local(addr, object_array, probe, exit_wave, auxiliary_wave, step)
+        POUK.pr_norm_local(addr_dev, probe_dev, probe_norm_dev)
+        POUK.ob_update_local(addr_dev, object_array_dev, probe_dev, exit_wave_dev, auxiliary_wave_dev, probe_norm_dev, A=0.5, B=0.5)
+        nPOUK.pr_norm_local(addr, probe, probe_norm)
+        nPOUK.ob_update_local(addr, object_array, probe, exit_wave, auxiliary_wave, probe_norm, A=0.5, B=0.5)
 
         np.testing.assert_allclose(object_array_dev.get(), object_array, rtol=1e-6, atol=1e-6,
                                       err_msg="The object array has not been updated as expected")
@@ -747,8 +750,6 @@ class PoUpdateKernelTest(PyCudaTest):
         total_number_modes = G * D
         A = total_number_scan_positions * total_number_modes # this is a 16 point scan pattern (4x4 grid) over all the modes
 
-        step = 0.5
-
         probe = np.empty(shape=(D, E, F), dtype=COMPLEX_TYPE)
         for idx in range(D):
             probe[idx] = np.ones((E, F)) * (idx + 1) + 1j * np.ones((E, F)) * (idx + 1)
@@ -761,6 +762,8 @@ class PoUpdateKernelTest(PyCudaTest):
         for idx in range(A):
             exit_wave[idx] = np.ones((B, C)) * (idx + 1) + 1j * np.ones((B, C)) * (idx + 1)
         auxiliary_wave = exit_wave.copy() * 1.5
+
+        object_norm = np.empty(shape=(1,B,C), dtype=FLOAT_TYPE)
 
         X, Y = np.meshgrid(range(scan_pts), range(scan_pts))
         X = X.reshape((total_number_scan_positions))
@@ -794,10 +797,13 @@ class PoUpdateKernelTest(PyCudaTest):
         probe_dev = gpuarray.to_gpu(probe)
         exit_wave_dev = gpuarray.to_gpu(exit_wave)
         auxiliary_wave_dev = gpuarray.to_gpu(auxiliary_wave)
+        object_norm_dev = gpuarray.to_gpu(object_norm)
         addr_dev = gpuarray.to_gpu(addr)
 
-        POUK.pr_update_local(addr_dev,  probe_dev, object_array_dev,exit_wave_dev, auxiliary_wave_dev, step)
-        nPOUK.pr_update_local(addr, probe, object_array, exit_wave, auxiliary_wave, step)
+        POUK.ob_norm_local(addr_dev, object_array_dev, object_norm_dev)
+        POUK.pr_update_local(addr_dev,  probe_dev, object_array_dev,exit_wave_dev, auxiliary_wave_dev, object_norm_dev, A=0.5, B=0.5)
+        nPOUK.ob_norm_local(addr, object_array, object_norm)
+        nPOUK.pr_update_local(addr, probe, object_array, exit_wave, auxiliary_wave, object_norm, A=0.5, B=0.5)
 
         np.testing.assert_allclose(probe_dev.get(), probe, rtol=1e-6, atol=1e-6,
                                       err_msg="The probe has not been updated as expected")
